requirements.txt
fastapi
uvicorn
python-pptx
streamlit
requests
python-dotenv
google-generativeai
pillow
